Create imutable.sol
The smart contract uses the immutable keyword to declare and initialize two immutable state variables in the constructor

'address public immutable MY_ADDRESS;' - This declares an immutable public state variable named 'MY_ADDRESS' of type 'address'. It is uninitialized in the contract body but will be initialized in the constructor.

'uint public immutable MY_UINT;' - This declares an immutable public state variable named 'MY_UINT' of type 'uint' (unsigned integer). Like MY_ADDRESS, it is uninitialized in the contract body but will be initialized in the constructor.

The 'constructor(uint _myUint)' is the constructor function that takes a parameter '_myUint' of type 'uint'. Inside the constructor, 'msg.sender' is assigned to 'MY_ADDRESS', and the input parameter '_myUint' is assigned to 'MY_UINT'. Once these values are assigned in the constructor, they cannot be changed.