Create ReadWriteStateVariable.sol
This is a basic smart contract

'uint public num;': This is a state variable of type 'uint' (unsigned integer) that is publicly accessible. It is used to store a number.

'function set(uint _num) public { num = _num; }': This function allows you to set the value of the state variable 'num'. It takes an argument '_num' and assigns its value to 'num'. This function requires a transaction to be sent to the blockchain because it modifies the state.

'function get() public view returns (uint) { return num; }': This function allows you to retrieve the value of the state variable 'num' without modifying the state. It is marked as 'view', indicating that it does not change the state and can be called without sending a transaction.