Create Constructor.sol
This code illustrates different ways to initialize parent contracts with parameters using inheritance.

Inheritance Initialization in Contract B:
In this case, the child contract 'B' directly specifies the parameters for the base contracts 'X' and 'Y' in the inheritance list. This way, each base contract is initialized with the provided parameters.

Inheritance Initialization in Contract C:
In this case, the child contract 'C' initializes the base contracts 'X' and 'Y' with parameters in its own constructor. The parameters are passed explicitly to the base contract constructors.

Order of Constructors in Contract D:
In this case, the child contract 'D' initializes the base contracts 'X' and 'Y' with parameters in its own constructor. The order of constructor calls is determined by the order of inheritance.

Order of Constructors in Contract E:
In this case, the child contract 'E' initializes the base contracts 'X' and 'Y' with parameters in its own constructor. The order of constructor calls is determined by the order of inheritance, regardless of the order specified in the child contract's constructor.