Create TryCatch.sol
The provided Solidity contract, Bar, demonstrates the usage of the try/catch mechanism with external calls and contract creation.

Foo Contract:

The 'Foo' contract is a simple contract used as an external contract in the 'Bar' contract.

Bar Contract:

The 'Bar' contract has two functions:

'tryCatchExternalCall': It makes an external call to the 'myFunc' function of the 'foo' contract using try/catch. It emits a 'log' message based on the result of the call.

'tryCatchNewContract': It creates a new 'Foo' contract using try/catch. It emits log messages based on the success or failure of the contract creation. If the creation is successful, it also provides access to the created 'Foo' contract through the variable 'foo'.